antonrx_backend/requirements.txt
fastapi==0.135.0
uvicorn==0.42.0
python-jose==3.5.0
cryptography==45.0.0
passlib==1.7.4
bcrypt==5.0.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic-extra-types==2.12.1
python-multipart==0.0.10
aiofiles==23.2.1
PyJWT==2.10.1

# API & Web
httpx==0.28.1
requests==2.32.3

# Database
firebase-admin==6.6.0
supabase==1.3.8
sqlalchemy==2.0.36
psycopg2-binary==2.9.12
pgvector==0.3.0

# Document Parsing
pypdf==5.3.1
pdfplumber==0.11.2
python-docx==1.0.2
beautifulsoup4==4.12.3
lxml==5.2.2
Pillow==10.3.0
pytesseract==0.3.13

# AI/ML
openai==1.13.3
anthropic==0.45.1

# Utilities
python-dotenv==1.0.1
python-slugify==8.0.1
email-validator==2.2.0

# Development (optional)
pytest==7.4.4
pytest-cov==4.1.0
black==24.3.0
flake8==7.1.1
mypy==1.11.0

# Logging
python-json-logger==2.0.9
